refactor(scripting): move workflow examples to workflows/ directory
Move deployFullChain, deployNewChain, initializeTokenBridge, and
transferOwnership from examples/ to workflows/ to reflect that they are
runnable end-to-end workflows rather than single-call demos. Update
commands.ts import paths accordingly. No content changes.
